refactor: 모듈을 하나로 합친다.
